@@ -1,6 +1,6 @@
 module jinya-fonts
 
-go 1.17
+go 1.19
 
 require (
 	github.com/gosimple/slug v1.12.0
